# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Restaurant Documentation",
        "description": "L5 Swagger OpenApi description for restaurant project",
        "contact": {
            "email": "drrifatalashwad0@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/v1.0/garage-times": {
            "patch": {
                "tags": [
                    "garage_times_management"
                ],
                "summary": "This method is to update garage times",
                "description": "This method is to update garage times",
                "operationId": "updateGarageTimes",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "garage_id",
                                    "times"
                                ],
                                "properties": {
                                    "garage_id": {
                                        "type": "number",
                                        "format": "number",
                                        "example": "1"
                                    },
                                    "times": {
                                        "type": "string",
                                        "format": "array",
                                        "example": [
                                            {
                                                "day": 0,
                                                "opening_time": "10:10",
                                                "closing_time": "10:15"
                                            },
                                            {
                                                "day": 1,
                                                "opening_time": "10:10",
                                                "closing_time": "10:15"
                                            },
                                            {
                                                "day": 2,
                                                "opening_time": "10:10",
                                                "closing_time": "10:15"
                                            },
                                            {
                                                "day": 3,
                                                "opening_time": "10:10",
                                                "closing_time": "10:15"
                                            },
                                            {
                                                "day": 4,
                                                "opening_time": "10:10",
                                                "closing_time": "10:15"
                                            },
                                            {
                                                "day": 5,
                                                "opening_time": "10:10",
                                                "closing_time": "10:15"
                                            },
                                            {
                                                "day": 6,
                                                "opening_time": "10:10",
                                                "closing_time": "10:15"
                                            }
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocesseble Content",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "not found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/v1.0/garage-times/{garage_id}": {
            "get": {
                "tags": [
                    "garage_times_management"
                ],
                "summary": "This method is to get garage times ",
                "description": "This method is to get garage times",
                "operationId": "getGarageTimes",
                "parameters": [
                    {
                        "name": "garage_id",
                        "in": "path",
                        "description": "garage_id",
                        "required": true,
                        "example": "6"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocesseble Content",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "not found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}